Moved character and weapon costs to their own files
