+++ src/types/WeaponCosts.ts
@@ -1,0 +1,19 @@
+export type WeaponCosts = {
+    name: string,
+    costs: {
+        mora: number,
+        wep_xp1: number,
+        wep_xp2: number,
+        wep_xp3: number,
+        ascension1: number,
+        ascension2: number,
+        ascension3: number,
+        ascension4: number,
+        elite1: number,
+        elite2: number,
+        elite3: number,
+        common1: number,
+        common2: number,
+        common3: number
+    }
+}
